+++ src/components/common/top-bar.tsx
@@ -73,8 +73,8 @@ export function TopBar() {
                 <div className={styles['header__search-menu-container']}>
                     <ul>
                         <li><a href='/'>Home</a></li>
-                        <li><a href='#'>Categories</a></li>
-                        <li><a href='#'>Shops</a></li>
+                        <li><a href='/categories'>Categories</a></li>
+                        <li><a href='/shops'>Shops</a></li>
                     </ul>
                 </div>
             </div>